+++ HISTORY.rst
@@ -2,11 +2,11 @@
 History
 =======
 
+* Reload when static assets or Jinja templates change.
+
 * Mention django-browser-reload in reload debug message.
 
 * Add Jinja template tag.
-
-* Reload when Jinja templates change.
 
 1.0.0 (2021-12-15)
 ------------------